refactor: extract status snapshot and normalization from registerCommands
Moves collectStatusSnapshot (179 lines), all normalize* validation
functions, and shared helpers into statusSnapshot.ts (406 lines).
registerCommands.ts reduced from 1405→1022 lines, now focused on
command registration wiring and UI handlers.
